+++ ecmean/performance_indices.py
@@ -22,13 +22,13 @@ from ecmean.libs.general import weight_split, get_domain, \
     check_time_axis, init_mydict, check_var_interface, check_var_climatology, \
     set_multiprocessing_start_method
 from ecmean.libs.files import var_is_there, get_inifiles, load_yaml, \
-    make_input_filename, get_clim_files
+    make_input_filename, get_clim_files, load_output_yaml
 from ecmean.libs.formula import formula_wrapper
 from ecmean.libs.masks import mask_field, select_region
 from ecmean.libs.areas import guess_bounds
 from ecmean.libs.units import units_extra_definition
 from ecmean.libs.ncfixers import xr_preproc, adjust_clim_file
-from ecmean.libs.plotting import heatmap_comparison_pi, prepare_clim_dictionaries_pi, plot_xarray
+from ecmean.libs.ecplotter import ECPlotter
 from ecmean.libs.parser import parse_arguments
 from ecmean.libs.loggy import setup_logger
 
@@ -187,53 +187,96 @@ class PerformanceIndices:
         # dump the yaml file for PI, including all the seasons (need to copy to avoid mess)
         if yamlfile is None:
             yamlfile = self.diag.filenames('yml')
+        self.loggy.info('Storing the performance indices in %s', yamlfile)
         with open(yamlfile, 'w', encoding='utf-8') as file:
             yaml.safe_dump(self.varstat, file, default_flow_style=False, sort_keys=False)
         self.toc('Storing')
 
-    def plot(self, mapfile=None, figformat='pdf'):
+
+
+    def plot(self, diagname='performance_indices', mapfile=None, figformat='pdf', storefig=True, returnfig=False):     
         """
         Generate the heatmap for performance indices.
 
         Args:
+            diagname (str): Name of the diagnostic. Default is 'performance_indices'.
             mapfile (str): Path to the output file. If None, it will be defined automatically following ECmean syntax.
-            figformat (str): Format of the output file. Default is 'pdf'.
+            storefig (bool): If True, store the figure in the specified file. Default is True.
+            returnfig (bool): If True, return the figure object. Default is False.
         """
-        if self.extrafigure:
-            self.loggy.debug('Plotting extra results...')
-            print(self.outarray)
-            debugfig = os.path.join(self.diag.figdir, f"map_{os.path.basename(self.diag.filenames('png'))}")
-            plot_xarray(self.outarray['map'], filename=debugfig, log_scale=True, cmap='viridis')
-            debugfig = os.path.join(self.diag.figdir, f"bias_{os.path.basename(self.diag.filenames('png'))}")
-            plot_xarray(self.outarray['bias'], filename=debugfig, cmap='seismic', log_scale=False)
-
-        # to this date, only EC23/EC24 support comparison with CMIP6 data
-        if self.diag.climatology in ['EC23', 'EC24']:
-
-            # load yaml file if is missing
-            if not self.varstat:
-                yamlfile = self.diag.filenames('yml')
-                self.loggy.info('Loading the stored data from the yaml file %s', yamlfile)
-                if os.path.isfile(yamlfile):
-                    with open(yamlfile, 'r', encoding='utf-8') as file:
-                        self.varstat = yaml.safe_load(file)
-                else:
-                    raise FileNotFoundError(f'File {yamlfile} not found.')
-
-            # prepare the data for the heatmap from the original yaml dictionaries
-            data2plot, cmip6, longnames = prepare_clim_dictionaries_pi(data=self.varstat,
-                                                                       clim=self.piclim,
-                                                                       shortnames=self.diag.field_all)
-
-            # call the heatmap routine for a plot
-            if mapfile is None:
-                mapfile = self.diag.filenames(figformat)
-
-            heatmap_comparison_pi(data_dict=data2plot, cmip6_dict=cmip6, diag=self.diag, longnames=longnames, filemap=mapfile)
-        else:
-            self.loggy.warning('Only EC23 and EC24 climatology is supported for comparison with CMIP6 data.')
-
+        plotter = ECPlotter(
+            diagnostic=diagname, modelname=self.diag.modelname,
+            expname=self.diag.expname, year1=self.diag.year1,
+            year2=self.diag.year2, regions=self.diag.regions,
+            seasons=self.diag.seasons)
+        if self.varstat is None:
+            self.varstat = load_output_yaml(self.diag.filenames('yml'))
+        if mapfile is None:
+            mapfile = self.diag.filenames(figformat)
+        fig = plotter.heatmap_plot(
+            data=self.varstat, reference=self.piclim,
+            variables=self.diag.field_all, climatology=self.diag.climatology,
+            filename=mapfile, storefig=storefig)
+        
         self.toc('Plotting')
+
+        if returnfig:
+            self.loggy.info('Returning figure object')
+            return fig
+        
+    # def plot(self, mapfile=None, figformat='pdf', returnfig=False, storefig=True):
+    #     """
+    #     Generate the heatmap for performance indices.
+
+    #     Args:
+    #         mapfile (str): Path to the output file. If None, it will be defined automatically following ECmean syntax.
+    #         figformat (str): Format of the output file. Default is 'pdf'.
+    #         returnfig (bool): If True, return the figure object. Default is False.
+    #         storefig (bool): If True, store the figure in the specified file. Default is True.
+    #     """
+    #     if self.extrafigure:
+    #         self.loggy.debug('Plotting extra results...')
+    #         print(self.outarray)
+    #         debugfig = os.path.join(self.diag.figdir, f"map_{os.path.basename(self.diag.filenames('png'))}")
+    #         plot_xarray(self.outarray['map'], filename=debugfig, log_scale=True, cmap='viridis')
+    #         debugfig = os.path.join(self.diag.figdir, f"bias_{os.path.basename(self.diag.filenames('png'))}")
+    #         plot_xarray(self.outarray['bias'], filename=debugfig, cmap='seismic', log_scale=False)
+
+    #     # to this date, only EC23/EC24 support comparison with CMIP6 data
+    #     if self.diag.climatology in ['EC23', 'EC24']:
+
+    #         # load yaml file if is missing
+    #         if not self.varstat:
+    #             yamlfile = self.diag.filenames('yml')
+    #             self.loggy.info('Loading the stored data from the yaml file %s', yamlfile)
+    #             if os.path.isfile(yamlfile):
+    #                 with open(yamlfile, 'r', encoding='utf-8') as file:
+    #                     self.varstat = yaml.safe_load(file)
+    #             else:
+    #                 raise FileNotFoundError(f'File {yamlfile} not found.')
+
+    #         # prepare the data for the heatmap from the original yaml dictionaries
+    #         self.loggy.debug('%s, %s, %s', self.varstat, self.piclim, self.diag.field_all)
+    #         data2plot, cmip6, longnames = prepare_clim_dictionaries_pi(data=self.varstat,
+    #                                                                    clim=self.piclim,
+    #                                                                    shortnames=self.diag.field_all)
+
+    #         # call the heatmap routine for a plot
+    #         if mapfile is None:
+    #             mapfile = self.diag.filenames(figformat)
+
+    #         fig = heatmap_comparison_pi(
+    #             data_dict=data2plot, cmip6_dict=cmip6, 
+    #             diag=self.diag, longnames=longnames, filemap=mapfile,
+    #             storefig=storefig)
+
+    #         self.toc('Plotting')
+    #         if returnfig:
+    #             return fig
+        
+    #     else:
+    #         self.loggy.warning('Only EC23 and EC24 climatology is supported for comparison with CMIP6 data.')
+
 
     @staticmethod
     def pi_worker(util, piclim, face, diag, field_3d, varstat, dictarray, varlist):