+++ tests/test_performance_indices.py
@@ -37,7 +37,7 @@ def test_performance_indices_cpld(clim):
 def test_performance_indices_amip(clim):
     performance_indices('amip', 1990, 1990, numproc=1,
                         climatology=clim, config='tests/config.yml', outputdir='tests/pluto')
-    file1 = 'tests/pluto/YAML/PI4_' + clim + '_amip_EC-Earth4_r1i1p1f1_1990_1990.yml'
+    file1 = 'tests/pluto/yml/PI4_' + clim + '_amip_EC-Earth4_r1i1p1f1_1990_1990.yml'
     file2 = 'tests/table/PI4_' + clim + '_amip_1990_1990.ref'
     with open(file1, 'r', encoding='utf8') as f1, open(file2, 'r', encoding='utf8') as f2:
         data1 = yaml.safe_load(f1)
@@ -47,7 +47,7 @@ def test_performance_indices_amip(clim):
 def test_performance_indices_omip():
     performance_indices('omip', 1990, 1990, numproc=1,
                         climatology='EC23', config='tests/config.yml', outputdir='tests/pluto')
-    file1 = 'tests/pluto/YAML/PI4_' + 'EC23' + '_omip_EC-Earth4_r1i1p1f1_1990_1990.yml'
+    file1 = 'tests/pluto/yml/PI4_' + 'EC23' + '_omip_EC-Earth4_r1i1p1f1_1990_1990.yml'
     file2 = 'tests/table/PI4_' + 'EC23' + '_omip_1990_1990.ref'
     with open(file1, 'r', encoding='utf8') as f1, open(file2, 'r', encoding='utf8') as f2:
         data1 = yaml.safe_load(f1)
@@ -59,7 +59,7 @@ def test_performance_indices_omip():
 @pytest.mark.parametrize("clim", ['EC23', 'EC24'])
 def test_cmd_performance_indices_CMIP6(clim):
     subprocess.run(['performance_indices', 'historical', '1990', '1990', '-j', '2', '-c',
-                    'tests/config_CMIP6.yml', '-k', clim, '-m', 'EC-Earth3', '-v', 'debug'],
+                    'tests/config_CMIP6.yml', '--climatology', clim, '-m', 'EC-Earth3', '-l', 'debug'],
                    env=env, check=True)
     file1 = 'tests/table/PI4_' + clim + '_historical_EC-Earth3_r1i1p1f1_1990_1990.yml'
     file2 = 'tests/table/PI4_' + clim + '_CMIP6_1990_1990.ref'